Fixup: delete rls_archive directory
Archeology has been limited, but I think the old rls workspace was
renamed to rls_archive, simultaneously the UI renamed rls to RLS, then
the gitauto script failed to reconcile the merge so I recreated the
rls_archive directory to fix it up -- but did so by copying the RLS
directory, without considering the duplicate ids this caused.

By pure chance we got away with this for the most part, until recent
mysterious crashes.  My theory is that geoserver would load the first
workspace it found, which happily was normally the correct RLS one,
but when perhaps a timestamp changed it was normally loading
rls_archive, then when a style referred to that id was referenced it
couldn't be found (in rls_archive not RLS), causing the exceptions.